move GenerationHyperparameters to realhf.api.core.model_api
