+++ realhf/impl/model/backend/inference.py
@@ -6,7 +6,6 @@ import transformers
 
 import realhf.api.core.model_api as model_api
 import realhf.base.constants as constants
-from realhf.experiments.common.gen_exp import GenerationHyperparameters
 from realhf.impl.model.backend.pipe_runner import PipelineRunner
 from realhf.impl.model.nn.real_llm_api import ReaLModel
 
@@ -92,8 +91,8 @@ class PipelinableInferenceEngine:
         packed_input_ids: torch.Tensor,
         cu_seqlens: torch.Tensor,
         tokenizer: transformers.PreTrainedTokenizerFast,
-        gconfig: GenerationHyperparameters = dataclasses.field(
-            default_factory=GenerationHyperparameters
+        gconfig: model_api.GenerationHyperparameters = dataclasses.field(
+            default_factory=model_api.GenerationHyperparameters
         ),
         num_micro_batches: Optional[int] = None,
     ):
